3 tests with super subtle bug fix
